# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LaunchCore SEO Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /todos/{id}:
    patch:
      operationId: updateTodo
      tags: [todos]
      summary: Update todo status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeoTodoUpdate"
      responses:
        "200":
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeoTodo"
        "404":
          description: Not found
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: searchLeads
      tags: [leads]
      summary: Start a lead search job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadSearchInput"
      responses:
        "201":
          description: Lead search started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadSearchJob"
...
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
...
components:
  schemas:
...
    SeoTodo:
      type: object
      required: [id, crawlId, title, priority, category, status, createdAt]
      properties:
        id:
          type: integer
        crawlId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        targetUrl:
          type: ["string", "null"]
        priority:
          type: string
          enum: [P1, P2, P3, P4]
        category:
          type: string
        exactAction:
          type: ["string", "null"]
        whyItMatters:
          type: ["string", "null"]
        expectedImpact:
          type: ["string", "null"]
        estimatedTimeMinutes:
          type: ["integer", "null"]
        implementationNotes:
          type: ["string", "null"]
        validationMethod:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, planned, in_progress, implemented, validated, ignored]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    SeoTodoUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, planned, in_progress, implemented, validated, ignored]
        implementationNotes:
          type: ["string", "null"]
    Report:
      type: object
      required: [id, crawlId, createdAt]
      properties:
        id:
          type: integer
        crawlId:
          type: integer
        executiveSummary:
          type: ["string", "null"]
        biggestRevenueLeaks:
          type: ["string", "null"]
        quickWins:
          type: ["string", "null"]
        technicalSeo:
          type: ["string", "null"]
        localSeo:
          type: ["string", "null"]
        conversionIssues:
          type: ["string", "null"]
        mobileSpeedReadiness:
          type: ["string", "null"]
        competitorGap:
          type: ["string", "null"]
        actionPlan30Days:
          type: ["string", "null"]
        outreachEmail:
          type: ["string", "null"]
        overallScore:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    LeadSearchInput:
      type: object
      required: [businessCategory, city]
      properties:
        businessCategory:
          type: string
        city:
          type: string
        country:
          type: string
          default: CA
        limit:
          type: integer
          default: 20
    LeadSearchJob:
      type: object
      required: [id, businessCategory, city, status, createdAt]
      properties:
        id:
          type: integer
        businessCategory:
          type: string
        city:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        leadsFound:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
    Lead:
      type: object
      required: [id, businessName, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        businessName:
          type: string
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        leadScore:
          type: ["integer", "null"]
        status:
          type: string
          enum: [new, contacted, qualified, disqualified, converted]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    LeadUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, qualified, disqualified, converted]
        notes:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
...